// src/types/page.ts
export interface BasePageConfig {
    type: 'about' | 'publication' | 'card' | 'text' | 'news';
    title: string;
    description?: string;
}

export interface PublicationPageConfig extends BasePageConfig {
    type: 'publication';
    source: string;
}

export interface TextPageConfig extends BasePageConfig {
    type: 'text';
    source: string;
}

export interface CardItem {
    title: string;
    subtitle?: string;
    date?: string;
    content?: string;
    tags?: string[];
    link?: string;
    image?: string;
}

export interface CardPageConfig extends BasePageConfig {
    type: 'card';
    items: CardItem[];
}

export interface NewsItem {
    date: string;
    content: string;
}

export interface NewsPageConfig extends BasePageConfig {
    type: 'news';
    source: string;
}